requirements.txt
assertpy==1.0
allure-pytest==2.6.2
